refactor: added border to buttons and merged branch
