+++ src/utils/base.scss
@@ -8,6 +8,7 @@ $secondary-color-1: #ff7043;
 $secondary-color-2: #bf360c;
 $secondary-color-3: #ff5722;
 $secondary-color-4: #ff8a65;
+
 $neutral-color-1: #121212;
 $neutral-color-2: #424242;
 $neutral-color-3: #9e9e9e;
